% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_project_member.R
\name{add_project_member}
\alias{add_project_member}
\title{Add a new member to a project}
\usage{
add_project_member(id = NULL, new_member = NULL)
}
\arguments{
\item{id}{the project id.}

\item{new_member}{the GBIF user to be added.}
}
\value{
A `list` of project information.
}
\description{
Add a new member to a project
}
