% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ruleset_member.R
\name{add_ruleset_member}
\alias{add_ruleset_member}
\title{Add a new member to a ruleset}
\usage{
add_ruleset_member(id = NULL, new_member = NULL)
}
\arguments{
\item{id}{the ruleset id.}

\item{new_member}{the GBIF user to be added.}
}
\value{
A `list` of ruleset information.
}
\description{
Add a new member to a ruleset
}
\examples{
\dontrun{
add_ruleset_member(1,"JOHN")
}
}
