% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_download.R
\name{clean_download}
\alias{clean_download}
\title{Clean a GBIF download of supicious records}
\usage{
clean_download(
  d,
  rm_suspicious = TRUE,
  rm_higher_order_suspicious = TRUE,
  handle_conflicts = "favor_supicious"
)
}
\arguments{
\item{d}{a GBIF download in DWCA format}

\item{rm_suspicious}{removes records with suspicious speciesKey}

\item{rm_higher_order_suspicious}{removes records with suspicious higher order keys}
}
\value{
A cleaned download `data.frame()`
}
\description{
Removes records that have been marked as suspicious by annotation users.
}
\examples{
\dontrun{
library(rgbif)
# Calopteryx 
# 0051416-241126133413365
# simple polygons not inverted 
# 0049858-241126133413365
# inverted polygon 
# 0049850-241126133413365
d <- occ_download_get('0051416-241126133413365') \%>\%
 occ_download_import()

clean_download(d) 
}


}
