% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_project.R
\name{delete_project}
\alias{delete_project}
\title{Delete a project}
\usage{
delete_project(id)
}
\arguments{
\item{id}{the id of the project}
}
\value{
list of information about deleted project
}
\description{
Delete a project
}
\examples{
\dontrun{
delete_project(1)
}
}
