% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_rule.R
\name{delete_rule}
\alias{delete_rule}
\title{Delete a rule}
\usage{
delete_rule(id)
}
\arguments{
\item{id}{the id of the rule}
}
\value{
list of information about deleted rule.
}
\description{
Delete a rule
}
