% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbifid_rule.R
\name{gbifid_rule}
\alias{gbifid_rule}
\title{Make an annotations using gbifid}
\usage{
gbifid_rule(
  gbifId = NULL,
  rulesetId = NULL,
  projectId = NULL,
  buffer = 1000,
  annotation = "SUSPICIOUS",
  create_rule = TRUE
)
}
\arguments{
\item{gbifId}{gbifId}

\item{rulesetId}{rulesetId}

\item{projectId}{projectId}

\item{buffer}{how much to buffer the bound box by.}

\item{annotation}{annoations}

\item{create_rule}{debugging delogical asking whether rule should be 
written to database.}
}
\value{
A `list` of info about the rule.
}
\description{
Make an annotations using gbifid
}
\details{
It isn't possible to attach annotations directly to gbifIds. However, it is 
possible to create a rule that captures almost the same thing by making a small 
bounding box around the lat-lon of the occurrence and indicating the 
datasetKey. This will accomplish close to same thing as attaching the 
annotations to the gbifId.
}
\examples{
\dontrun{
gbifid_rule(1986620884,projectId = 1,rulesetId=1,create_rule=FALSE)
}
}
