% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project.R
\name{get_project}
\alias{get_project}
\title{Get a project}
\usage{
get_project(id = NULL, offset = NULL, limit = NULL)
}
\arguments{
\item{id}{the id of the project.}

\item{offset}{page start.}

\item{limit}{number of records to return on page.}
}
\value{
a `tibble`.
}
\description{
Get a project
}
