% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rule.R
\name{get_rule}
\alias{get_rule}
\title{Get rules}
\usage{
get_rule(id = NULL, limit = NULL, offset = NULL, ...)
}
\arguments{
\item{id}{rule id.}

\item{limit}{page start.}

\item{offset}{number of records to return on page.}

\item{...}{Additional args passed to query.}
}
\value{
a tibble of rules
}
\description{
Get rules
}
\examples{
\dontrun{
get_rule()
}
}
