% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rule_comment.R
\name{get_rule_comment}
\alias{get_rule_comment}
\title{Get comments on rule}
\usage{
get_rule_comment(id = NULL)
}
\arguments{
\item{id}{the id of the rule}
}
\value{
a `tibble` of comments.
}
\description{
Get comments on rule
}
\examples{
\dontrun{
get_rule_comment(1)
}
}
