% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ruleset.R
\name{get_ruleset}
\alias{get_ruleset}
\title{Get a ruleset.}
\usage{
get_ruleset(id = NULL, projectId = NULL, limit = NULL, offset = NULL)
}
\arguments{
\item{id}{the id of the ruleset.}

\item{projectId}{id of project.}

\item{limit}{page start.}

\item{offset}{number of records to return on page.}
}
\description{
Get a ruleset.
}
