% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rule.R
\name{make_rule}
\alias{make_rule}
\title{Make a rule}
\usage{
make_rule(...)
}
\arguments{
\item{...}{A named list of parameters (geometry and annotation required).}
}
\value{
A list of information about the rule.
}
\description{
Make a rule
}
\details{
Annotation needs to be part of controlled vocab.
}
