% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_contest_rule.R
\name{rm_contest_rule}
\alias{rm_contest_rule}
\title{Remove contest of a rule}
\usage{
rm_contest_rule(id)
}
\arguments{
\item{id}{id of rule to contest.}
}
\value{
A `list` information about the rule contested.
}
\description{
Remove contest of a rule
}
\examples{
\dontrun{
rm_contest_rule(1)
}
}
