% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_support_rule.R
\name{rm_support_rule}
\alias{rm_support_rule}
\title{Remove support for a rule}
\usage{
rm_support_rule(id = NULL)
}
\arguments{
\item{id}{the id of the rule to updated.}
}
\value{
A `list` information about the rule supported.
}
\description{
Remove support for a rule
}
