% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_rule.R
\name{support_rule}
\alias{support_rule}
\title{Support a rule}
\usage{
support_rule(id = NULL)
}
\arguments{
\item{id}{the id of the rule to updated.}
}
\value{
A `list` information about the rule supported.
}
\description{
Support a rule
}
