% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_project.R
\name{update_project}
\alias{update_project}
\title{Update a project}
\usage{
update_project(
  id = NULL,
  name = NULL,
  description = NULL,
  members = NULL,
  deleted = NULL,
  createdBy = NULL,
  keep_members = TRUE
)
}
\arguments{
\item{id}{project id.}

\item{name}{new name of project.}

\item{description}{new description for project.}

\item{members}{new members. If `keep_members=TRUE`, then new will be added to old members.}

\item{deleted}{logical should the project be deleted or restored.}

\item{createdBy}{new creator of project.}

\item{keep_members}{keep old members. Default is TRUE.}
}
\value{
A `list` with data of the new project.
}
\description{
Update a project
}
\details{
Update a project. If fields are left `NULL`, then they will not be updated. 
If `keep_memebers=TRUE`, new members are added to the list of old members.
}
\examples{
\dontrun{
update_project(1,name="New name")
}
}
