% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_ruleset.R
\name{update_ruleset}
\alias{update_ruleset}
\title{Update a ruleset}
\usage{
update_ruleset(
  id = NULL,
  projectId = NULL,
  name = NULL,
  description = NULL,
  members = NULL,
  deleted = NULL,
  createdBy = NULL,
  keep_members = TRUE
)
}
\arguments{
\item{id}{of ruleset.}

\item{name}{new name of ruleset.}

\item{description}{new description of ruleset.}

\item{members}{new members of ruleset. If `keep_members=TRUE`, then new will be added to old members.}

\item{deleted}{logical should the ruleset be deleted or restored.}

\item{createdBy}{new creator of ruleset.}

\item{keep_members}{keep old members. Default is TRUE.}
}
\value{
A `list` of an updated ruleset.
}
\description{
Update a ruleset
}
