% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rule.R
\name{make_rule}
\alias{make_rule}
\title{Make a rule}
\usage{
make_rule(taxonKey = NULL, geometry = NULL, annotation = NULL, ...)
}
\arguments{
\item{taxonKey}{GBIF taxonKey for which rule applies to.}

\item{geometry}{WKT text of the rule.}

\item{annotation}{One of the controlled vocabulary.}

\item{...}{A named list of parameters (geometry and annotation required).}
}
\value{
A `list` of information about the rule.
}
\description{
Make a rule
}
\details{
Annotation needs to be part of controlled vocab.
}
