% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rule_comment.R
\name{make_rule_comment}
\alias{make_rule_comment}
\title{Make a comment on rule.}
\usage{
make_rule_comment(id = NULL, comment = NULL)
}
\arguments{
\item{comment}{comment.}

\item{ruleId}{id of the rule you want to leave a comment on.}
}
\description{
Make a comment on rule.
}
